requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
chromadb==0.5.15
langchain==0.3.1
langchain-community==0.3.1
langchain-anthropic==0.2.3
langchain-text-splitters==0.3.0
sentence-transformers==3.2.1
numpy>=1.26.0
streamlit==1.40.0
pandas==2.2.3
plotly==5.24.1
pymupdf==1.24.10
python-docx==1.1.2
openai==1.54.0
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
loguru==0.7.2
